All storage tests are passing
